+++ tests/testthat/test-star_database.R
@@ -481,17 +481,17 @@ test_that("set_dimension_attribute_names() and get_dimension_attribute_names()",
   })
 })
 
-test_that("set_measure_names() and get_measure_names()", {
+test_that("set_fact_measure_names() and get_fact_measure_names()", {
   expect_equal({
     star_database(mrs_cause_schema, ft_num) |>
-      set_measure_names(
+      set_fact_measure_names(
         measures = c(
           "Pneumonia and Influenza",
           "All",
           "Rows Aggregated"
         )
       ) |>
-      get_measure_names()
+      get_fact_measure_names()
   }, {
     c("Pneumonia and Influenza", "All", "Rows Aggregated")
   })