+++ NEWS.md
@@ -1,6 +1,6 @@
 # rolap 2.1.0
 * Definition of role-playing and role dimensions (with explanatory vignette): `role_playing_dimension()` and `get_role_playing_dimension_names()` functions.
-* New functions to query and rename attributes in dimensions and measures in facts: `get_dimension_attribute_names()`, `set_dimension_attribute_names()`, `get_measure_names()` and `set_measure_names()`.
+* New functions to query and rename attributes in dimensions and measures in facts: `get_dimension_attribute_names()`, `set_dimension_attribute_names()`, `get_fact_measure_names()` and `set_fact_measure_names()`.
 * Include detailed validation error messages.
 
 # rolap 2.0.0